refactor(0.3.0): remove usize from models

+++ contracts/src/models/resources.cairo
@@ -14,7 +14,7 @@ struct ResourceCost {
     #[key]
     entity_id: u128,
     #[key]
-    index: usize,
+    index: u32,
     resource_type: u8,
     amount: u128
 }
